overhaul logging behavior
Instead of a single logger in a daqling instance, logging is now
separated into 'core' and 'module'. When a log entry is made in the
loaded module, it is logged via a module logger instance. All other
entires are logged through a core logger instance. Different levels can
be applied on these two logger instances. The same logging API applies.

Presently, DaqlingCore is built with external linkage for
daqutils::Logger::m_logger in mind; thus binaries it is linked to must
define this symbol.